+++ HISTORY.md
@@ -18,6 +18,7 @@
 
 ### 2022-03-18
 * objects, gui: refactored sprite serialization, added SpriteMorph >> toXMLString() api
+* blocks: added optional receiver sprite to toXMLScript() api
 
 ### 2022-03-17
 * blocks: refactored dependencies scan for scripts